Basic sketch working
here the size of the grid is hard coded to 16x16 and i have orange background and borders on the cells so they can be identified. I set the attributes of the cells when added via adding a class. tried to do via div.style.cssText and could not get it working; I want to figure out why; I it will be easier adding the programable nature when this is set in js rather than css.

I had to expand the size of my main continer by 2px to account for the border that i added. otherwise i could not get a full row of 16 cells per row.